Fixed issue with starting LOVE from GUI
