+++ C#/Event-ECS-WPF/Controls/ModulesControl.xaml.cs
@@ -1,4 +1,4 @@
-﻿using Event_ECS_WPF.Projects;
+﻿using Event_ECS_WPF.Projects.Love;
 using System.Windows;
 using System.Windows.Controls;
