+++ C#/Event-ECS-WPF/Lua/conf.lua
@@ -1,0 +1,45 @@
+function love.conf(t)
+    t.accelerometerjoystick = {0}      -- Enable the accelerometer on iOS and Android by exposing it as a Joystick(boolean)
+    t.externalstorage = {1}           -- True to save files(and read from the save directory) in external storage on Android(boolean)
+    t.gammacorrect = {2}             -- Enable gamma-correct rendering, when supported by the system(boolean)
+
+    t.audio.mixwithsystem = {3}       -- Keep background music playing when opening LOVE(boolean, iOS and Android only)
+
+    t.window.width = {4}                -- The window width(number)
+    t.window.height = {5}               -- The window height(number)
+    t.window.borderless = {6}         -- Remove all border visuals from the window(boolean)
+    t.window.resizable = {7}          -- Let the window be user-resizable(boolean)
+    t.window.minwidth = {8}               -- Minimum window width if the window is resizable(number)
+    t.window.minheight = {9}             -- Minimum window height if the window is resizable(number)
+    t.window.fullscreen = {10}         -- Enable fullscreen(boolean)
+    t.window.fullscreentype = '{11}' -- Choose between 'desktop' fullscreen or 'exclusive' fullscreen mode(string)
+    t.window.vsync = {12}             -- Vertical sync mode(number)
+    t.window.msaa = {13}                  -- The number of samples to use with multi-sampled antialiasing(number)
+    t.window.display = {14}                -- Index of the monitor to show the window in (number)
+    t.window.highdpi = {15}            -- Enable high-dpi mode for the window on a Retina display (boolean)
+
+    t.modules.audio = {16}-- Enable the audio module (boolean)
+    t.modules.data = {17}-- Enable the data module (boolean)
+    t.modules.event = {18}              -- Enable the event module (boolean)
+    t.modules.font = {19}               -- Enable the font module (boolean)
+    t.modules.graphics = {20}-- Enable the graphics module (boolean)
+    t.modules.image = {21}-- Enable the image module (boolean)
+    t.modules.joystick = {22}-- Enable the joystick module (boolean)
+    t.modules.keyboard = {23}-- Enable the keyboard module (boolean)
+    t.modules.math = {24}-- Enable the math module (boolean)
+    t.modules.mouse = {25}-- Enable the mouse module (boolean)
+    t.modules.physics = {26}-- Enable the physics module (boolean)
+    t.modules.sound = {27}-- Enable the sound module (boolean)
+    t.modules.system = {28}-- Enable the system module (boolean)
+    t.modules.thread = {29}-- Enable the thread module (boolean)
+    t.modules.timer = {30}-- Enable the timer module (boolean), Disabling it will result 0 delta time in love.update
+    t.modules.touch = {31}-- Enable the touch module (boolean)
+    t.modules.video = {32}-- Enable the video module (boolean)
+    t.modules.window = {33}-- Enable the window module (boolean)
+
+    t.identity = '{34}'	-- The name of the save directory
+	t.window.title = '{34}'	-- The window title (string)
+    t.appendidentity = false -- Search files in source directory before save directory (boolean)
+    t.version = '11.0' -- The LOVE version this game was made for (string)
+    t.console = false -- Attach a console (boolean, WIndows only)
+end
